requirements.txt
scikit-learn
pandas
numpy
matplotlib
minio
pyyaml
openpyxl